Basic UCGI implementation

+++ core/src/xoshiro.c
@@ -57,7 +57,7 @@ void xoshiro_jump(XoshiroPRNG *prng) {
   uint64_t s1 = 0;
   uint64_t s2 = 0;
   uint64_t s3 = 0;
-  for (int i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
+  for (unsigned long i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
     for (int b = 0; b < 64; b++) {
       if (JUMP[i] & UINT64_C(1) << b) {
         s0 ^= prng->s[0];
@@ -87,7 +87,7 @@ void xoshiro_long_jump(XoshiroPRNG *prng) {
   uint64_t s1 = 0;
   uint64_t s2 = 0;
   uint64_t s3 = 0;
-  for (int i = 0; i < sizeof LONG_JUMP / sizeof *LONG_JUMP; i++)
+  for (unsigned long i = 0; i < sizeof LONG_JUMP / sizeof *LONG_JUMP; i++)
     for (int b = 0; b < 64; b++) {
       if (LONG_JUMP[i] & UINT64_C(1) << b) {
         s0 ^= prng->s[0];
